Added Discounting framework. Basket now contains a discounter object, this tracks discounts as items are added to the Basket
An "Offer" is a wrapper for a discount mechanism
A "Discount" is a specific instance of a discount - the idea being that
cases such as meal deals can be dealt with by a per meal discount.

For the momemt, an Offer applies to a single product, but this is not intended to be the final result.